move some declarations to shared file
